separate modal component from button
